backend/requirements.txt
Django==4.2.4
Pillow==10.0.0
autopep8==2.0.2
bcrypt==4.0.1
django-cors-headers==4.2.0
django-colorfield==0.9.0
django-pgtrigger==4.7.0
django-smart-selects==1.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
whitenoise==6.5.0